1.修复多个寄存器 写入错误的Bug

+++ android/src/main/java/com/reactlibrary/RNModbusModule.java
@@ -127,7 +127,7 @@ public class RNModbusModule extends ReactContextBaseJavaModule {
     short[] cmd = new short[length];
     for (int i = 0; i < length; i++) {
       int number =  values.getInt(i);
-      cmd[i] =(byte)number ;
+      cmd[i] =(short)number ;
     }
     ModbusManager.get()
             .writeRegisters(slaveId, start, cmd,
